requirements.txt
pandas
polars
pyarrow
yfinance
Quantlib
scipy

# Web 
aiohttp
duckduckgo-search
googlesearch-python
requests 
beautifulsoup4

# Ollama 
ollama

# Other 
pyyaml